It now opens a window to select sprites.
there is no (current) way to remove a sprite.
typing:
load <url>
will load the image(on filesystem) into the editor, if more than one is
loaded, they will be in tabs.
typing:
xml
will cause it to dump the generated xml forms of the sprites, however id
and name are likely to change